+++ ElvUI_MerathilisUI/Core/Changelog/6.46.lua
@@ -1,8 +1,13 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[646] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "16.12.2024",
+	FIXES = {
+		"[KeyFeedback]: Should now work again",
+	},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[ObjectiveTracker]: Wowheadlinks should now redirect to the english comments",
+		"[Skins]: Updated various skins",
+	},
 }