+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Changes ###
 
-  * [Guild]: Added a workauround to let the guild message not spam around
-  * [OmniCD]: Sync some borders with the skin
-  * [Changelog]: Let the changelog show the display version
+  * [KeyFeedback]: Should now work again
+  * [ObjectiveTracker]: Wowheadlinks should now redirect to the english comments
+  * [Skins]: Updated various skins